--- a/hw10_program_optimization/go.mod
+++ b/hw10_program_optimization/go.mod
@@ -1,11 +1,16 @@
-module github.com/fixme_my_friend/hw10_program_optimization
+module github.com/realayyo/hw10_program_optimization
 
-go 1.19
-
-require github.com/stretchr/testify v1.7.0
+go 1.21
 
 require (
+	github.com/goccy/go-json v0.10.2
+	github.com/stretchr/testify v1.7.0
+)
+
+require (
+	github.com/aclements/go-moremath v0.0.0-20210112150236-f10218a38794 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	golang.org/x/perf v0.0.0-20231127181059-b53752263861 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b // indirect
 )