--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -1,11 +1,12 @@
 package hw10programoptimization
 
 import (
-	"encoding/json"
+	"bufio"
 	"fmt"
 	"io"
-	"regexp"
 	"strings"
+
+	json "github.com/goccy/go-json"
 )
 
 type User struct {
@@ -31,36 +32,35 @@ func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 type users [100_000]User
 
 func getUsers(r io.Reader) (result users, err error) {
-	content, err := io.ReadAll(r)
-	if err != nil {
-		return
-	}
+	scanner := bufio.NewScanner(r)
+	var user User
 
-	lines := strings.Split(string(content), "\n")
-	for i, line := range lines {
-		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
-			return
+	i := 0
+	for scanner.Scan() {
+		err := json.Unmarshal(scanner.Bytes(), &user)
+		if err != nil {
+			return result, fmt.Errorf("unmarshal error: %w", err)
 		}
 		result[i] = user
+		i++
 	}
+
 	return
 }
 
 func countDomains(u users, domain string) (DomainStat, error) {
 	result := make(DomainStat)
+	lowerDomain := strings.ToLower(domain)
 
 	for _, user := range u {
-		matched, err := regexp.Match("\\."+domain, []byte(user.Email))
-		if err != nil {
-			return nil, err
-		}
+		lowerEmail := strings.ToLower(user.Email)
+		atIndex := strings.LastIndex(lowerEmail, "@")
 
-		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
-			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+		if atIndex != -1 && strings.HasSuffix(lowerEmail[atIndex+1:], lowerDomain) {
+			domainPart := lowerEmail[atIndex+1:]
+			result[domainPart]++
 		}
 	}
+
 	return result, nil
 }